pyproject.toml
[project]
name = "gpx2exif"
dynamic = ["version"]
description = "Command-line tool to add GPS info from a GPX file to EXIF tags in images"
license = "GPL-3.0-or-later"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Guilhem Vellut", email = "g@vellut.com" }]
keywords = ["gpx", "gps", "exif", "tag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]
dependencies = [
    "click~=8.0",
    "click-config-file~=0.6.0",
    "colorama~=0.4.0",
    "addict~=2.4.0",
    "piexif~=1.1.0",
    "gpxpy~=1.6.0",
    "pandas~=2.3.0",
    "pytz",
    "simplekml~=1.3.0",
    "flickrapi~=2.4.0",
    "python-dateutil~=2.9.0",
]

[project.optional-dependencies]
vision = ["google-cloud-vision~=3.0"]

[tool.uv]
override-dependencies = [ # Annoying warnings/logs
    "grpcio==1.67.1",
]

[project.urls]
Homepage = "https://github.com/gvellut/gpx2exif"
Repository = "https://github.com/gvellut/gpx2exif"
Issues = "https://github.com/gvellut/gpx2exif/issues"
Source = "https://github.com/gvellut/gpx2exif"

[project.scripts]
gpx2exif = "gpx2exif.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "gpx2exif/__init__.py"

[tool.hatch.build.targets.sdist]
include = ["gpx2exif/"]
